generators/app/templates/requirements.txt
Django~=1.10.4
# django-ninecms>=0.6.0
coverage~=4.3.1
# django-admin-bootstrapped~=2.5.6
# django-admin-bootstrapped-plus>=0.1.1
# django-bootstrap3~=7.0.1
django-braces==1.10.0
django-debug-toolbar~=1.6.0
django-envelope==1.2
djangorestframework==3.5.3
mysqlclient~=1.3.7
newrelic~=2.78.0.57
python3-memcached~=1.51
sqlparse~=0.2.2
